+++ src/extension.js
@@ -1,7 +1,9 @@
+/* eslint-disable no-template-curly-in-string */
 /* eslint-disable import/no-unresolved */
 /* eslint-disable import/no-useless-path-segments */
 const vscode = require("vscode");
 const View = require("./view");
+const Snippet = require("./snippet");
 const SnippetsEditor = require("./snippets-editor");
 
 // this is included for webpack, so that it picks up the CSS file
@@ -18,11 +20,18 @@ function activate(context) {
       let snippetsEditor = new SnippetsEditor(context);
       let activeEditor = vscode.window.activeTextEditor;
 
-      let selectedText = "";
+      let bodyText = "${3}";
       if (activeEditor) {
-        selectedText = activeEditor.document.getText(activeEditor.selection);
+        let selectedText = activeEditor.document.getText(
+          activeEditor.selection
+        );
+        if (selectedText.length > 0) {
+          bodyText = selectedText;
+        }
       }
-      snippetsEditor.add(selectedText);
+
+      let snippet = new Snippet("${1:enter a name}", "${2}", bodyText, "${4}");
+      snippetsEditor.add(snippet);
     })
   );
 }