+++ src/snippets-editor.js
@@ -20,9 +20,9 @@ class SnippetsEditor {
   /**
    * @async
    * Add a new snippets to a user snippet file based on the language that the user selects.
-   * * @param {String} bodyText The body of the snippet you want to create.
+   * * @param {Snippet} snippet The snippet you want to add.
    */
-  async add(bodyText) {
+  async add(snippet) {
     let language = await util.askForLanguage();
 
     if (language === undefined) {
@@ -35,40 +35,8 @@ class SnippetsEditor {
       console.log(`Cannot open user snippets file. ${err.message}`);
     }
 
-    this.insertSnippet(bodyText);
-  }
-
-  /**
-   * @async
-   * Insert a new snippet to the current text editor.
-   * * @param {String} body The body of the snippet you want to create.
-   */
-  async insertSnippet(body) {
-    if (!this.editor) {
-      return;
-    }
-
-    let document = this.editor.document;
-    let eolChar = "\n";
-
-    if (document && document.eol === vscode.EndOfLine.CRLF) {
-      eolChar = "\r\n";
-    }
-
-    let snippetString =
-      eolChar + '"${1:enter a name}": {' + eolChar + ' "prefix": "${2}",';
-
-    if (body.length > 1) {
-      let bodyArray = formatter.convertToArray(body, eolChar);
-      snippetString += `${eolChar} "body":  ${JSON.stringify(bodyArray)},`;
-    } else {
-      snippetString += eolChar + ' "body": ["${3}"],';
-    }
-
-    snippetString +=
-      eolChar + ' "description" : "${4}"' + eolChar + "},${0}" + eolChar;
-
     let range = new vscode.Range(0, 1, 0, 1);
+    let snippetString = `${snippet.toString()},`;
     this.editor.insertSnippet(new vscode.SnippetString(snippetString), range);
   }
 }